feat: create analysis app
